ramdisk

+++ src/Plugin/Function_ramdisk_speedtest.php
@@ -26,14 +26,16 @@ function function_ramdisk_speedtest(Parse $parse, Data $data){
     $config = $object->data_read();
     if($config){
         $url = $config->get('ramdisk.url');
-        $command = 'dd if=/dev/zero of=' . $url . 'zero bs=4k count=100000';
-        Core::execute($object, $command, $output);
-        echo 'Write:' . PHP_EOL;
-        echo $output . PHP_EOL;
-        $command = 'dd if=' . $url . 'zero of=/dev/null bs=4k count=100000';
-        Core::execute($object, $command, $output);
-        Echo 'Read:' . PHP_EOL;
-        echo $output . PHP_EOL;
-        File::delete($url . 'zero');
+        if($url){
+            $command = 'dd if=/dev/zero of=' . $url . 'zero bs=4k count=100000';
+            Core::execute($object, $command, $output);
+            echo 'Write:' . PHP_EOL;
+            echo $output . PHP_EOL;
+            $command = 'dd if=' . $url . 'zero of=/dev/null bs=4k count=100000';
+            Core::execute($object, $command, $output);
+            Echo 'Read:' . PHP_EOL;
+            echo $output . PHP_EOL;
+            File::delete($url . 'zero');
+        }
     }
 }
